Update telling-git-about-your-signing-key.md
Users should be told to point the signing key to a file instead of commiting the duplication of a file into config.
I also experienced an error in windows powershell terminal that the temporary file could not be created when signing, as a result of using the copy-paste method. Setting the user.signingKey to ~\.ssh\id_rsa.pub resolved the error.

I'd have done/modified a code example, but the reusable code chunks aren't in this file and no idea where they are.